Move detection files to sub-folder
