+++ cellfinder_napari/detect/__init__.py
@@ -1,0 +1,1 @@
+from .detect import detect
