+++ cellfinder_napari/detect/input_containers.py
